OfflineRAG_Pro/requirements.txt

gradio>=4.44.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
torch>=2.0.0
pillow>=10.0.0
pytesseract>=0.3.10
python-docx>=1.1.0
PyPDF2>=3.0.0
openai-whisper>=20231117
requests>=2.31.0
numpy>=1.24.0
